# scrivia-v2-openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scrivia API v2 - LLM Tools",
    "description": "API Scrivia v2 pour l'intégration LLM avec support complet des notes, dossiers et classeurs. Tous les endpoints supportent les slugs et UUIDs.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://scrivia.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Note": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "source_title": { "type": "string", "maxLength": 255 },
          "markdown_content": { "type": "string" },
          "html_content": { "type": "string" },
          "header_image": { "type": "string", "format": "uri" },
          "folder_id": { "type": "string", "format": "uuid" },
          "classeur_id": { "type": "string", "format": "uuid" },
          "user_id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "description": { "type": "string", "maxLength": 500 },
          "visibility": { "type": "string", "enum": ["private", "shared", "members", "public"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "source_title", "user_id", "created_at", "updated_at"]
      },
...
      "PublishNotePayload": {
        "type": "object",
        "properties": {
          "visibility": { "type": "string", "enum": ["private", "shared", "members", "public"] }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "integer" }
        },
        "required": ["error"]
      }
    }
  },
  "paths": {
...
    "/api/v2/note/{ref}/metadata": {
      "get": {
        "operationId": "getNoteMetadata",
        "summary": "Récupérer les métadonnées d'une note",
        "description": "Récupérer les informations de base d'une note (titre, dates, etc.)",
        "tags": ["Notes"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug de la note"
          }
        ],
        "responses": {
          "200": {
            "description": "Métadonnées récupérées avec succès",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Note" }
              }
            }
          },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
    "/api/v2/note/{ref}/publish": {
      "post": {
        "operationId": "publishNote",
        "summary": "Publier une note",
        "description": "Changer le statut de publication d'une note",
        "tags": ["Notes"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug de la note"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PublishNotePayload" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note publiée avec succès",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Note" }
              }
            }
          },
          "400": { "$ref": "#/components/schemas/ErrorResponse" },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
    "/api/v2/note/{ref}/statistics": {
      "get": {
        "operationId": "getNoteStatistics",
        "summary": "Récupérer les statistiques d'une note",
        "description": "Récupérer les statistiques détaillées d'une note (nombre de mots, temps de lecture, etc.)",
        "tags": ["Notes"],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID ou slug de la note"
          }
        ],
        "responses": {
          "200": {
            "description": "Statistiques récupérées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "word_count": { "type": "integer" },
                    "character_count": { "type": "integer" },
                    "line_count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/schemas/ErrorResponse" }
        }
      }
    },
...
  }
} 